+++ lib/src/migrations/migrator.dart
@@ -4,8 +4,6 @@ import 'dart:async';
 import 'dart:io'; // For Directory and File access
 import 'package:path/path.dart' as p; // For path manipulation
 import 'package:eloquent/eloquent.dart'; // Main package imports
-import 'migration_repository_interface.dart';
-import 'migration.dart';
 //import '../support/str.dart'; 
 
 /// Manages and runs database migrations.