+++ lib/src/schema/grammars/schema_mysql_grammar.dart
@@ -1,9 +1,5 @@
 //schema_mysql_grammar.dart
 import 'package:eloquent/eloquent.dart';
-import 'package:eloquent/src/doctrine/schema/foreign_key_constraint.dart';
-import 'package:eloquent/src/doctrine/schema/index.dart';
-import 'package:eloquent/src/doctrine/schema/table_diff.dart';
-import 'package:eloquent/src/doctrine/schema/unique_constraint.dart';
 import 'package:meta/meta.dart';
 
 /// Schema Grammar specific to MySQL.